Refactoring and new mock data
