--- .dart_tool/chrome-device/Default/Extension State/LOG
@@ -1,3 +1,0 @@
-2023/03/15-20:19:03.161 22c Reusing MANIFEST C:\Users\amanx\AppData\Local\Temp\flutter_tools.3d7e6fe\flutter_tools_chrome_device.59f400fb\Default\Extension State/MANIFEST-000001
-2023/03/15-20:19:03.162 22c Recovering log #3
-2023/03/15-20:19:03.162 22c Reusing old log C:\Users\amanx\AppData\Local\Temp\flutter_tools.3d7e6fe\flutter_tools_chrome_device.59f400fb\Default\Extension State/000003.log
