--- .dart_tool/chrome-device/Default/Extension State/LOG.old
@@ -1,3 +1,0 @@
-2023/03/15-15:18:37.498 5640 Reusing MANIFEST C:\Users\amanx\AppData\Local\Temp\flutter_tools.abcdcf3d\flutter_tools_chrome_device.cd812c0\Default\Extension State/MANIFEST-000001
-2023/03/15-15:18:37.499 5640 Recovering log #3
-2023/03/15-15:18:37.500 5640 Reusing old log C:\Users\amanx\AppData\Local\Temp\flutter_tools.abcdcf3d\flutter_tools_chrome_device.cd812c0\Default\Extension State/000003.log
