+++ .dart_tool/chrome-device/Default/Network/Network Persistent State
@@ -1,1 +1,1 @@
-{"net":{"http_server_properties":{"servers":[{"alternative_service":[{"advertised_alpns":["h3"],"expiration":"13325973537045729","port":443,"protocol_str":"quic"}],"anonymization":[],"network_stats":{"srtt":15572},"server":"https://optimizationguide-pa.googleapis.com","supports_spdy":true},{"anonymization":[],"server":"https://unpkg.com","supports_spdy":true},{"alternative_service":[{"advertised_alpns":["h3"],"expiration":"13325991543582805","port":443,"protocol_str":"quic"}],"anonymization":[],"network_stats":{"srtt":29760},"server":"https://accounts.google.com","supports_spdy":true},{"alternative_service":[{"advertised_alpns":["h3"],"expiration":"13325991551438622","port":443,"protocol_str":"quic"}],"anonymization":[],"network_stats":{"srtt":29185},"server":"https://fonts.gstatic.com","supports_spdy":true}],"supports_quic":{"address":"2600:1700:73b0:9e0:10f5:ac74:c49c:e2a1","used_quic":true},"version":5},"network_qualities":{"CAASABiAgICA+P////8B":"4G"}}}
+{"net":{"http_server_properties":{"servers":[{"alternative_service":[{"advertised_alpns":["h3"],"expiration":"13325973537045729","port":443,"protocol_str":"quic"}],"anonymization":[],"network_stats":{"srtt":15572},"server":"https://optimizationguide-pa.googleapis.com","supports_spdy":true},{"alternative_service":[{"advertised_alpns":["h3"],"expiration":"13326132079550261","port":443,"protocol_str":"quic"}],"anonymization":[],"network_stats":{"srtt":29760},"server":"https://accounts.google.com","supports_spdy":true},{"anonymization":[],"server":"https://unpkg.com","supports_spdy":true},{"alternative_service":[{"advertised_alpns":["h3"],"expiration":"13326132083600744","port":443,"protocol_str":"quic"}],"anonymization":[],"network_stats":{"srtt":30524},"server":"https://fonts.gstatic.com","supports_spdy":true}],"supports_quic":{"address":"2600:1700:73b0:9e0:2590:3cf:cbbb:7ebf","used_quic":true},"version":5},"network_qualities":{"CAASABiAgICA+P////8B":"4G","CAISABiAgICA+P////8B":"4G"}}}
