+++ .dart_tool/chrome-device/Default/Service Worker/Database/LOG
@@ -1,3 +1,2 @@
-2023/03/15-20:19:03.099 b14 Reusing MANIFEST C:\Users\amanx\AppData\Local\Temp\flutter_tools.3d7e6fe\flutter_tools_chrome_device.59f400fb\Default\Service Worker\Database/MANIFEST-000001
-2023/03/15-20:19:03.100 b14 Recovering log #3
-2023/03/15-20:19:03.100 b14 Reusing old log C:\Users\amanx\AppData\Local\Temp\flutter_tools.3d7e6fe\flutter_tools_chrome_device.59f400fb\Default\Service Worker\Database/000003.log 
+2023/03/17-11:22:15.457 49b8 Creating DB C:\Users\amanx\AppData\Local\Temp\flutter_tools.79d79d66\flutter_tools_chrome_device.f27dfaa6\Default\Service Worker\Database since it was missing.
+2023/03/17-11:22:15.462 49b8 Reusing MANIFEST C:\Users\amanx\AppData\Local\Temp\flutter_tools.79d79d66\flutter_tools_chrome_device.f27dfaa6\Default\Service Worker\Database/MANIFEST-000001
