+++ .dart_tool/chrome-device/Default/Site Characteristics Database/LOG
@@ -1,3 +1,4 @@
-2023/03/15-20:19:02.870 1ff4 Reusing MANIFEST C:\Users\amanx\AppData\Local\Temp\flutter_tools.3d7e6fe\flutter_tools_chrome_device.59f400fb\Default\Site Characteristics Database/MANIFEST-000001
-2023/03/15-20:19:02.871 1ff4 Recovering log #3
-2023/03/15-20:19:02.871 1ff4 Reusing old log C:\Users\amanx\AppData\Local\Temp\flutter_tools.3d7e6fe\flutter_tools_chrome_device.59f400fb\Default\Site Characteristics Database/000003.log 
+2023/03/17-11:21:19.152 56f8 Reusing MANIFEST C:\Users\amanx\AppData\Local\Temp\flutter_tools.79d79d66\flutter_tools_chrome_device.f27dfaa6\Default\Site Characteristics Database/MANIFEST-000001
+2023/03/17-11:21:19.153 56f8 Recovering log #6
+2023/03/17-11:21:19.153 56f8 Reusing old log C:\Users\amanx\AppData\Local\Temp\flutter_tools.79d79d66\flutter_tools_chrome_device.f27dfaa6\Default\Site Characteristics Database/000006.log 
+2023/03/17-11:21:19.153 56f8 Delete type=0 #3
