--- .dart_tool/chrome-device/Default/shared_proto_db/metadata/LOG
@@ -1,3 +1,0 @@
-2023/03/15-20:19:03.103 6ee8 Reusing MANIFEST C:\Users\amanx\AppData\Local\Temp\flutter_tools.3d7e6fe\flutter_tools_chrome_device.59f400fb\Default\shared_proto_db\metadata/MANIFEST-000001
-2023/03/15-20:19:03.104 6ee8 Recovering log #3
-2023/03/15-20:19:03.105 6ee8 Reusing old log C:\Users\amanx\AppData\Local\Temp\flutter_tools.3d7e6fe\flutter_tools_chrome_device.59f400fb\Default\shared_proto_db\metadata/000003.log
