--- .dart_tool/chrome-device/Default/shared_proto_db/metadata/LOG.old
@@ -1,3 +1,0 @@
-2023/03/15-15:18:37.461 6ce0 Reusing MANIFEST C:\Users\amanx\AppData\Local\Temp\flutter_tools.abcdcf3d\flutter_tools_chrome_device.cd812c0\Default\shared_proto_db\metadata/MANIFEST-000001
-2023/03/15-15:18:37.462 6ce0 Recovering log #3
-2023/03/15-15:18:37.462 6ce0 Reusing old log C:\Users\amanx\AppData\Local\Temp\flutter_tools.abcdcf3d\flutter_tools_chrome_device.cd812c0\Default\shared_proto_db\metadata/000003.log
