+++ .dart_tool/package_config_subset
@@ -1,3 +1,7 @@
+namer_app
+2.18
+file:///C:/Personal%20Use/Programming/CIS4951/DriverDrowsinessDetection/
+file:///C:/Personal%20Use/Programming/CIS4951/DriverDrowsinessDetection/lib/
 sky_engine
 2.12
 file:///C:/Personal%20Use/Programming/CIS4951/flutter/bin/cache/pkg/sky_engine/
@@ -10,10 +14,6 @@ flutter_test
 2.17
 file:///C:/Personal%20Use/Programming/CIS4951/flutter/packages/flutter_test/
 file:///C:/Personal%20Use/Programming/CIS4951/flutter/packages/flutter_test/lib/
-namer_app
-2.18
-file:///C:/Personal%20Use/Programming/CIS4951/flutterFrontEnd/DriverDrowsinessDetection/
-file:///C:/Personal%20Use/Programming/CIS4951/flutterFrontEnd/DriverDrowsinessDetection/lib/
 async
 2.18
 file:///C:/Users/amanx/AppData/Local/Pub/Cache/hosted/pub.dev/async-2.10.0/
